requirements.txt
torch
transformers
accelerate
chronos-forecasting

pandas
numpy
scikit-learn
matplotlib
seaborn

python-dateutil
tqdm
requests
huggingface_hub[hf_xet]
pytest